backend/requirements.txt
attrs==25.3.0
audioread==3.0.1
basic-pitch==0.4.0
cattrs==25.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
contourpy==1.3.2
coremltools==8.3.0
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.0
fonttools==4.60.0
idna==3.10
joblib==1.5.2
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.0
matplotlib==3.10.6
mido==1.3.3
mir_eval==0.8.2
mpmath==1.3.0
msgpack==1.1.1
noisereduce==3.0.3
numba==0.62.0
numpy==2.2.6
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
pretty_midi==0.2.10
protobuf==6.32.1
pyaml==25.7.0
PyAudio==0.2.14
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
resampy==0.4.2
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Flask==3.0.3
Flask-CORS==5.0.0

python-dotenv==1.0.0
